lib: add a broken implementation of extend_replace_indices
One of the tests is failing. I haven't written all the tests yet, so future
tests will probably fail too. Has to do with an array-out-of-bounds access.

This is what happens when the human body only gets six hours of sleep. At least
I remembered to write tests.

I wonder if it's possible to leverage iterators to get rid of indices. That
usually makes code a lot nicer to deal with.