lib: make more work towards coherent functional extension function
This isn't necessarily one single logical set of changes. I just want to save my
work so that when I mess up in the future I can come back to this point.

We've implemented in this commit a CopyInstructor iterator which instructs an
implementation how to functionally update a single circle.

Next we probably need to implement a recursive copy instructor that will tell an
implementation how to functionally update a circle and the circles it contains.