lib: add function that maps a Graph to a FlatGraph
Objects of the Graph type carry a whole lot of information: everything needed to
render the graph, as well as the complete history of everything done in the
system.

This is useful when traversing back and forth through time---since all the
information is always present---but it isn't great when we need to actually go
about rendering it on screen.

It turns out we don't need a lot of the information present in the Graph type
while rendering, so we discard it and create a FlatGraph instead. This type
only holds the groups visible in the current frame of time; it's much simpler
than Graph and therefore is easier (and quicker) to traverse.