Draw circles using triangles instead of fragment shader hackery
We used to draw a box using two triangles, sending one such box for each circle
instance, along with a matrix that scaled the box and translated it to the right
place. The vertex shader would apply this transformation to each vertex in the
box.

The vertex shader would:
- transform the box by the instance matrix
- calculate the midpoint by applying the instance matrix to (0,0,0)
- calculate the radius by (M*(0,0,0)-M*(1,0,0)).x, where M is the instance matrix
- send (midpoint, radius) to the fragment shader

The fragment shader would:
- receive an interpolated position within the box
- receive (midpoint, radius)
- (*) determine the distance from the interpolated position to the radius
- output a color whose alpha value is based upon (*)

This approach produces nicely antialiased circles. Unfortunately, it seems a bit
difficult to reason about when we want to compensate for non-square windows.

The way to do this is to create a perspective projection that squashes the
coordinate system in the largest screen dimension, so that when the normalized
device coordinates are mapped to the window, they'll be "unstretched" into a
square.

In our previous system, this means that each box we send to the fragment shader
will no longer be a square---each will be a rectangle. Our fragment shader then
has to figure out the distance of the interpolated point not from the radius of
the circle, but from the circumference of the ellipse that fits inside the
rectangle. I couldn't find an easy way to do this that would be fast enough to
bother doing on every single fragment.

So instead of sending boxes, we just send tessellated circles. Each "circle" is
composed of a bunch of triangles. There is only one such tessellation; each
circle is just an instance of that tessellation coupled with a view matrix.

Both our vertex and fragment shaders do a whole lot less work now, since they
don't have to figure out the radius of the circle or the distance from the
interpolated point to its radius.

Upsides:
- perspective transformations are easier to deal with
- aspect ratio scaling is easier (possible) to deal with

Downsides:
- Circles are no longer antialiased

Unknowns:
- might be faster, since the fragment shader is doing less work
- might be slower, because we're using way more vertices