Add group()
This is half of our basis for proper modification of FlatGraphs. group() lets us
add nodes in any way we want.

The missing piece---for when we make a mistake when adding nodes---is remove(),
which does what you would expect.

This commit also makes the assumption that no branch in a FlatGraph can have
less than two indications.