Build and display text based on graph content
The program is much slower now when run with opt-level < 2 (i.e., in a
non-release mode). Building with --release produces the expected fast zooming /
panning at runtime.

I'm not sure yet if this is due to an inefficiency in the code or something
else.

Unfortunately, --release builds take about twice as long as normal builds, which
already take ~ 12 seconds, at least on my computer.