Only display indications, not inclusions of the selected circle.
Before this commit, left-clicking on an indication of the currently-selected
enclosure would zoom it into the middle of the screen. The (now previously)-selected
enclosure would be zoomed as well so that it still encloses the indication that
was clicked on.

After this commit, left-clicking on an indication of the currently-selected
enclosure still zooms it into the middle of the screen. The difference is that
the (now previously)-selected enclosure is not displayed at all.

This change is meant to address several issues.

First, it is not clear to me in the old system which enclosure should be
displayed around the currently-selected one. If our enclosures form a tree, we
have no problems. But if they are not a tree and are a DAG (or worse, a cyclic
graph), there may be multiple such enclosures that could reasonably be displayed
around the currently selected enclosure. With this change, we simply do not
display any of them.

Second: how, in the old system, should we decide which enclosure to display
around the first node we select? In the naming example (naming, vowel, consonant
groups), the first node we select is included in itself. This was evident if you
zoom in on the naming group far enough. But when the application first starts
up, we didn't display anything surrounding it. This seemed like not the right
solution, because both nodes are the *same*, so they should be displayed *in the
same way*. But, of course, there could be multiple valid enclosures to display
around the currently selected one---so we end up with our first problem. The
only reason we can display the surrounding enclosures in the old case when
zooming is because we kept track of the history of selected nodes, and just
displayed those.

Third: the old system made calculating the coordinates of circles harder. Not
impossible, but harder. My naive implementation drew the first circle with a
radius of 1.0. When we zoomed in, the newly focused circle would be drawn with a
smaller radius than 1.0. Unfortunately, we run into floating-point error when we
zoom in too far. So we would have to transition to a system where we always (or
most of the time) draw the focused node with a radius of 1.0. But now we have to
figure out what size the nodes that enclose the current node should be, which
means we have to either reverse the equations we already have, or just start
rendering from radius > 1.0 and work back down to our focused node. All of this
didn't seem to me like a good, simple solution. Plus, it would end up being
computationally more complicated and more inefficient.

The new system is a whole lot easier to reason about and is more efficient. So
I'm going to stick with it for now.