Instead of storing data in graph, use a table and put keys in graph
I'm about to add image support. The main graph-modification algorithm (enclose)
needs to be able to clone nodes sometimes. It would be unfortunate if this would
involve cloning an entire image bitmap.

So instead of having the graph own the data, we have a 'store::Store' own the
data. Each graph node that originally had a string now has a u64 key which it
can pass to the Store to get the string.

The Store uses a hash map from u64 keys to String (and eventually image) values.
The keys are computed as the hash of the values. So "hello world" is stored
under the key hash-value-of("hello world").

Why not just use a HashSet? Why the extra indirection of computing the hash of
the hash involved with HashMaps? -- because we can't have the FlatGraph data
type hold references, it has to own its data. If it held references, the
lifetimes would propagate up into the State data structure, which does not seem
like a good idea.

Another reason - when serializing the state of the application (not yet
implemented), we will probably want to make the file name of the serialized
stuff the hash of the serialized stuff.