Calculate text scale without binary search
It turns out that the aspect ratio of text stays (essentially) the same no
matter what scale you render it at. Instead of binary searching for the text
scale, we can just calculate the scale once, then use the aspect ratio to figure
out exactly how big the text should be, and finally use the difference between
the current size and the desired size to get the proper text scale.